Time: 0 ms (100%), Space: 8.1 MB (92.7%) - LeetHub

+++ 1490-generate-a-string-with-characters-that-have-odd-counts/1490-generate-a-string-with-characters-that-have-odd-counts.cpp
@@ -9,7 +9,7 @@ public:
             --n;
         }
 
-        for (unsigned short i = 0; i < n; ++i)
+        for (int i = 0; i < n; ++i)
             answer.push_back('a');
 
         return answer;
